refactor: use barrel exports
