+++ jobs/Frontend/src/const/api.ts
@@ -1,0 +1,2 @@
+export const API_URL = 'https://api.themoviedb.org/3'
+export const API_KEY = import.meta.env.VITE_API_KEY
