+++ jobs/Frontend/src/const/index.ts
@@ -1,2 +1,1 @@
-export const API_URL = 'https://api.themoviedb.org/3'
-export const API_KEY = import.meta.env.VITE_API_KEY
+export * from './api'
